LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointerleave_descendants-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
pointerleave

Test Description: This test checks if pointerleave event works properly.
Put your mouse over the black rectangle and then move it out through purple rectangle boundaries.


Pointer Events pointerleave tests

The following pointer types were detected: (none).

Refresh the page to run the tests again with a different pointer type.


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN pointerleave event received
